Remove old NodeDetail component
