+++ src/es-table.js
@@ -3,80 +3,80 @@ import ESPanel from './es-panel';
 import numeral from 'numeral';
 
 class ESTable extends ESPanel {
-	columnHeader(column) {
-		if (column.props.hasOwnProperty('name')) return column.props.name;
+  columnHeader(column) {
+    if (column.props.hasOwnProperty('name')) return column.props.name;
 
-		return column.props.source.charAt(0).toUpperCase() + column.props.source.slice(1);
-	}
+    return column.props.source.charAt(0).toUpperCase() + column.props.source.slice(1);
+  }
 
-	getDescendantProp (obj, path) {
-		return path.split('.').reduce((acc, part) => acc && acc[part], obj);
-	}
+  getDescendantProp (obj, path) {
+    return path.split('.').reduce((acc, part) => acc && acc[part], obj);
+  }
 
-	columnValue(row, column) {
-		var format = column.props.format;
-		var source = column.props.source;
-		var value = this.getDescendantProp(row, source);
+  columnValue(row, column) {
+    var format = column.props.format;
+    var source = column.props.source;
+    var value = this.getDescendantProp(row, source);
 
-		switch (format) {
+    switch (format) {
 
-			case 'string':
-				return value;
-				break;
+      case 'string':
+        return value;
+        break;
 
-			case 'date':
-				return 'Momentum';
-				break;
+      case 'date':
+        return 'Momentum';
+        break;
 
-			case 'size':
-				return numeral(value).format('0.0b');
-				break;
+      case 'size':
+        return numeral(value).format('0.0b');
+        break;
 
-			case 'number':
-				return numeral(value).format('0.0a');
-				break;
-		}
-	}
+      case 'number':
+        return numeral(value).format('0.0a');
+        break;
+    }
+  }
 
-	getBody() {
-		var rows = this.getRows();
+  getBody() {
+    var rows = this.getRows();
 
-		rows = rows.map( (row) => {
-			return (
-				<tr key={row.name}>
-					{this.props.children.map( (column) => {
+    rows = rows.map( (row) => {
+      return (
+        <tr key={row.name}>
+          {this.props.children.map( (column) => {
 
-						return (
-							<td key={column.props.source}>
-								{this.columnValue(row, column)}
-							</td>
-						);
-					})}
-				</tr>
-			);
-		});
+            return (
+              <td key={column.props.source}>
+                {this.columnValue(row, column)}
+              </td>
+            );
+          })}
+        </tr>
+      );
+    });
 
-		if (rows.length == 0) rows = this.getAlternateBody();
+    if (rows.length == 0) rows = this.getAlternateBody();
 
-		return (
-			<table className="table table-responsive table-striped">
-		        <thead>
-		        <tr>
-		        	{this.props.children.map( (column) => {
-		        		return (<th key={column.props.source}>{this.columnHeader(column)}</th>);
-		        	})}
-		        </tr>
-		        </thead>
-		        <tbody>
-					{rows}
-		        </tbody>
-	      	</table>
-		);
-	}
+    return (
+      <table className="table table-responsive table-striped">
+            <thead>
+            <tr>
+              {this.props.children.map( (column) => {
+                return (<th key={column.props.source}>{this.columnHeader(column)}</th>);
+              })}
+            </tr>
+            </thead>
+            <tbody>
+          {rows}
+            </tbody>
+          </table>
+    );
+  }
 
-	getAlternateBody() {
-		return (<tr><td colSpan={React.Children.count(this.props.children)}>No data</td></tr>);
-	}
+  getAlternateBody() {
+    return (<tr><td colSpan={React.Children.count(this.props.children)}>No data</td></tr>);
+  }
 
 }
